+++ buildSrc/src/main/kotlin/Dependencies.kt
@@ -1,9 +1,8 @@
 // Version information for dependencies
 
 object Versions {
-    const val moshi = "1.9.3"
-    const val okio = "2.6.0"
-    const val ktor = "1.3.2"
+    const val moshi = "1.14.0"
+    const val ktor = "2.1.1"
     const val hamkrest = "1.7.0.3"
     const val junit = "5.6.2"
     const val logback = "1.2.3"
@@ -27,17 +26,17 @@ object Deps {
         const val reflection = "${Groups.moshi}:moshi-kotlin:${Versions.moshi}"
     }
     object Ktor {
+        const val serialization = "${Groups.ktor}:ktor-serialization:${Versions.ktor}"
         const val server = "${Groups.ktor}:ktor-server-core:${Versions.ktor}"
+        const val serverCallLogging = "${Groups.ktor}:ktor-server-call-logging:${Versions.ktor}"
+        const val serverContentNegotiation = "${Groups.ktor}:ktor-server-content-negotiation:${Versions.ktor}"
         const val clientCore = "${Groups.ktor}:ktor-client-core:${Versions.ktor}"
         const val clientCIO = "${Groups.ktor}:ktor-client-cio:${Versions.ktor}"
-        const val clientJson = "${Groups.ktor}:ktor-client-json:${Versions.ktor}"
-        const val clientJsonJvm = "${Groups.ktor}:ktor-client-json-jvm:${Versions.ktor}"
+        const val clientContentNegotiation = "${Groups.ktor}:ktor-client-content-negotiation:${Versions.ktor}"
         const val netty = "${Groups.ktor}:ktor-server-netty:${Versions.ktor}"
         const val testHost = "${Groups.ktor}:ktor-server-test-host:${Versions.ktor}"
         const val clientMock = "${Groups.ktor}:ktor-client-mock:${Versions.ktor}"
-        const val clientMockJvm = "${Groups.ktor}:ktor-client-mock-jvm:${Versions.ktor}"
     }
-    const val okio = "${Groups.okio}:okio:${Versions.okio}"
     const val logback = "${Groups.logback}:logback-classic:${Versions.logback}"
     const val hamkrest = "${Groups.hamkrest}:hamkrest:${Versions.hamkrest}"
     object Junit {