+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+* Migrate to Ktor 2.
+* Combine client and server libraries into a single `ktor-moshi` library with no client or server dependencies.
+
 ## [2.1.0] - 2020-07-18
 
 ### Added